build: update Storybook to v8
Update Storybook to v8. Remove incompatible addons
(@etchteam/storybook-addon-status) and its configuration.

Simplify config:

- Move the decorator to its own .tsx file and rename main.tsx and
  preview.tsx to main.ts and preview.ts respectively
- Add types

Rename *.stories.mdx -> *.mdx as this is now required for Storybook 8.

Remove unused vite.config.ts

Remove unused dependencies.

Remove Storybook deps from .ncurc.major.cjs
