+++ .ncurc.major.cjs
@@ -2,6 +2,6 @@ const minorConfig = require('./.ncurc.minor.cjs');
 
 module.exports = {
   ...minorConfig,
-  reject: [...minorConfig.reject, '@mdx-js/react', 'storybook', '@storybook/*', 'eslint'],
+  reject: [...minorConfig.reject, 'eslint'],
   target: 'latest',
 };
