+++ package.json
@@ -38,7 +38,6 @@
     "postcss": "8.4.38",
     "prettier": "3.2.5",
     "stylelint": "16.3.1",
-    "stylelint-config-prettier": "9.0.5",
     "stylelint-config-standard-scss": "13.1.0",
     "stylelint-order": "6.0.4",
     "typescript": "5.4.5"
