+++ packages/storybook/config/ParameterArgsDecorator.tsx
@@ -1,0 +1,12 @@
+import type { Decorator } from '@storybook/react';
+
+export const ParameterArgsDecorator: Decorator = (Story, context) => {
+  // Hack to make current args for a story available in the transformSource of the docs addon
+  context.parameters['args'] = context.args;
+
+  return (
+    <div className="voorbeeld-theme">
+      <Story />
+    </div>
+  );
+};
