+++ packages/storybook/config/main.ts
@@ -1,0 +1,27 @@
+import type { StorybookConfig } from '@storybook/react-vite';
+
+const config: StorybookConfig = {
+  stories: ['../src/**/*stories.@(js|jsx|ts|tsx)', '../src/**/*.mdx'],
+  addons: [
+    '@storybook/addon-a11y',
+    '@storybook/addon-actions',
+    '@storybook/addon-docs',
+    '@storybook/addon-interactions',
+    '@storybook/addon-links',
+    '@storybook/addon-viewport',
+  ],
+  framework: {
+    name: '@storybook/react-vite',
+    options: {},
+  },
+  core: {
+    disableTelemetry: true,
+    disableWhatsNewNotifications: true,
+  },
+  docs: {
+    autodocs: true,
+  },
+  staticDirs: ['../../../proprietary/assets/src'],
+};
+
+export default config;
