+++ packages/storybook/package.json
@@ -19,11 +19,6 @@
     "storybook": "storybook dev --config-dir config/ --port 6006"
   },
   "devDependencies": {
-    "@babel/core": "7.24.4",
-    "@babel/preset-env": "7.24.4",
-    "@babel/preset-react": "7.24.1",
-    "@babel/preset-typescript": "7.24.1",
-    "@etchteam/storybook-addon-status": "4.2.4",
     "@kernteam/assets": "workspace:*",
     "@kernteam/components-css": "workspace:*",
     "@kernteam/design-tokens": "workspace:*",
@@ -32,39 +27,27 @@
     "@kernteam/web-components-stencil": "workspace:*",
     "@nl-design-system-unstable/nlds-design-tokens": "1.0.0-alpha.189",
     "@nl-design-system-unstable/voorbeeld-design-tokens": "1.0.0-alpha.133",
-    "@storybook/addon-a11y": "7.6.17",
-    "@storybook/addon-actions": "7.6.17",
-    "@storybook/addon-docs": "7.6.17",
-    "@storybook/addon-interactions": "7.6.17",
-    "@storybook/addon-links": "7.6.17",
-    "@storybook/addon-viewport": "7.6.17",
-    "@storybook/manager-api": "7.6.17",
-    "@storybook/preset-scss": "1.0.3",
-    "@storybook/react": "7.6.17",
-    "@storybook/react-vite": "7.6.17",
-    "@storybook/react-webpack5": "7.6.17",
-    "@storybook/testing-library": "0.2.2",
-    "@storybook/theming": "7.6.17",
+    "@storybook/addon-a11y": "8.0.8",
+    "@storybook/addon-actions": "8.0.8",
+    "@storybook/addon-docs": "8.0.8",
+    "@storybook/addon-interactions": "8.0.8",
+    "@storybook/addon-links": "8.0.8",
+    "@storybook/addon-viewport": "8.0.8",
+    "@storybook/blocks": "8.0.8",
+    "@storybook/react": "8.0.8",
+    "@storybook/react-vite": "8.0.8",
+    "@storybook/theming": "8.0.8",
     "@tabler/icons-react": "3.2.0",
     "@types/react": "18.2.75",
     "@types/react-dom": "18.2.24",
     "@utrecht/component-library-react": "3.0.0",
     "@utrecht/web-component-library-react": "1.0.2",
-    "@vitejs/plugin-react": "4.2.1",
-    "babel-loader": "9.1.3",
-    "css-loader": "7.1.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "require-from-string": "2.0.2",
     "rimraf": "5.0.5",
-    "sass": "1.74.1",
-    "sass-loader": "14.1.1",
-    "storybook": "7.6.17",
-    "style-loader": "4.0.0",
+    "storybook": "8.0.8",
     "typescript": "5.4.5",
-    "vite": "5.2.8",
-    "vite-plugin-html": "3.2.2",
-    "vite-plugin-markdown": "2.2.0",
-    "webpack": "5.91.0"
+    "vite": "5.2.8"
   }
 }
