+++ packages/storybook/src/documentation/license.mdx
