+++ packages/storybook/src/documentation/notice.mdx
@@ -1,4 +1,4 @@
-import { Markdown, Meta } from "@storybook/addon-docs";
+import { Markdown, Meta } from "@storybook/blocks";
 import markdown from "../../../../NOTICE.md?raw";
 
 <Meta title="Example/Toestemming voor gebruik" />
