+++ packages/storybook/tsconfig.json
@@ -5,6 +5,6 @@
     "skipLibCheck": true
   },
   "extends": "../../tsconfig.json",
-  "include": ["config/**/*.ts", "config/**/*.tsx", "src/**/*.ts", "src/**/*.tsx", "vite.config.ts"],
+  "include": ["config/**/*.ts", "src/**/*.ts", "src/**/*.tsx"],
   "exclude": ["**/node_modules/*"]
 }
